+++ R/standalone.R
@@ -1,21 +1,33 @@
+#' Standalone MDT
+#'
+#' This function launches a standalone testing session for the MDT.
+#' This can be used for data collection, either in the laboratory or online.
+#' @param title (Scalar character) Title to display during testing.
+#' @param admin_password (Scalar character) Password for accessing the admin panel.
+#' @param researcher_email (Scalar character)
+#' If not \code{NULL}, this researcher's email address is displayed
+#' at the bottom of the screen so that online participants can ask for help.
+#' @param languages (Character vector)
+#' Determines the languages available to participants.
+#' Possible languages include English (\code{"EN"}), French (\code{"French"}),
+#' and German (\code{"DE"}).
+#' The first language is selected by default
+#' @param dict The psychTestR dictionary used for internationalisation.
+#' @param ... Further arguments to be passed to \code{\link{mdt}()}.
 #' @export
-standalone_mdt <- function(num_items = 20L,
-                           take_training = TRUE,
-                           feedback = mdt::mdt.feedback.no_score(),
-                           title = "Melody discrimination test",
+standalone_mdt <- function(title = "Melody discrimination test",
                            admin_password = "replace-with-secure-password",
-                           researcher_email = "p.m.c.harrison@qmul.ac.uk",
+                           researcher_email = NULL,
                            languages = mdt_languages(),
-                           dict = mdt::mdt_dict) {
+                           dict = mdt::mdt_dict,
+                           ...) {
   elts <- c(
     psychTestR::new_timeline(
       psychTestR::get_p_id(prompt = psychTestR::i18n("enter_p_id"),
                            button_text = psychTestR::i18n("AMDI_0016_I_0001_1")),
       dict = dict
     ),
-    mdt::mdt(num_items = num_items,
-             take_training = take_training,
-             feedback = feedback),
+    mdt::mdt(dict = dict, ...),
     psychTestR::elt_save_results_to_disk(complete = TRUE),
     psychTestR::new_timeline(
       psychTestR::final_page(shiny::p(